+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 - smelter type max temperature
 - smelter block registration api
 - mold registration api
+- clay and brick molds
 ### Fixed:
 - 
 ### Changed: